Add Sophonics LoRA Int5 DeepMLP non-record submission
