+++ records/track_non_record_16mb/2026-03-27_Sophonics_LoRA_Int5_DeepMLP/environment_5090.log
@@ -1,0 +1,29 @@
+Sun Apr 19 14:26:51 UTC 2026
+3e93ec78ff82
+origin	https://github.com/DesperateTomatoCultivator/parameter-golf.git (fetch)
+origin	https://github.com/DesperateTomatoCultivator/parameter-golf.git (push)
+495dd4a4599b97147b702d3914b2423d82ba54c5
+Sun Apr 19 14:26:51 2026       
++-----------------------------------------------------------------------------------------+
+| NVIDIA-SMI 570.195.03             Driver Version: 570.195.03     CUDA Version: 12.8     |
+|-----------------------------------------+------------------------+----------------------+
+| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
+| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
+|                                         |                        |               MIG M. |
+|=========================================+========================+======================|
+|   0  NVIDIA GeForce RTX 5090        On  |   00000000:1C:00.0 Off |                  N/A |
+|  0%   27C    P8             18W /  575W |       0MiB /  32607MiB |      0%      Default |
+|                                         |                        |                  N/A |
++-----------------------------------------+------------------------+----------------------+
+                                                                                         
++-----------------------------------------------------------------------------------------+
+| Processes:                                                                              |
+|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
+|        ID   ID                                                               Usage      |
+|=========================================================================================|
+|  No running processes found                                                             |
++-----------------------------------------------------------------------------------------+
+Python 3.12.3
+Traceback (most recent call last):
+  File "<stdin>", line 1, in <module>
+ModuleNotFoundError: No module named 'sentencepiece'
